# docs/openapi/agent-platform-openapi.json
{
  "components": {
    "schemas": {
...
      "ApiLabelSchemaListResponse": {
        "properties": {
          "include_archived": {
            "title": "Include Archived",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiLabelSchemaRecordResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "include_archived"
        ],
        "title": "ApiLabelSchemaListResponse",
        "type": "object"
      },
      "ApiLabelSchemaRecordResponse": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "preview": {
            "title": "Preview",
            "type": "string"
          },
          "schema": {
            "additionalProperties": true,
            "title": "Schema",
            "type": "object"
          },
          "source_path": {
            "title": "Source Path",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "description",
          "schema",
          "preview",
          "archived",
          "source_path",
          "updated_at"
        ],
        "title": "ApiLabelSchemaRecordResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LabelSchemaPatchRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema"
          }
        },
        "title": "LabelSchemaPatchRequest",
        "type": "object"
      },
      "LabelSchemaWriteRequest": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          },
          "schema": {
            "additionalProperties": true,
            "title": "Schema",
            "type": "object"
          }
        },
        "required": [
          "key",
          "schema"
        ],
        "title": "LabelSchemaWriteRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Provides versioned API routes for Agent Platform runtime/control/test orchestration. Designed for backend-first API consumption and ADE frontend integration.",
    "summary": "Runtime and control APIs for ADE and local Agent Platform workflows",
    "title": "Agent Platform API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/platform/schema-center/label-schemas": {
      "get": {
        "operationId": "api_schema_center_list_label_schemas_api_v1_platform_schema_center_label_schemas_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLabelSchemaListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Label Lab JSON schemas",
        "tags": [
          "Schema Center"
        ]
      },
      "post": {
        "operationId": "api_schema_center_create_label_schema_api_v1_platform_schema_center_label_schemas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelSchemaWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLabelSchemaRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Label Lab JSON schema",
        "tags": [
          "Schema Center"
        ]
      }
    },
    "/api/v1/platform/schema-center/label-schemas/{key}": {
      "get": {
        "operationId": "api_schema_center_get_label_schema_api_v1_platform_schema_center_label_schemas__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLabelSchemaRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Label Lab JSON schema",
        "tags": [
          "Schema Center"
        ]
      },
      "patch": {
        "operationId": "api_schema_center_update_label_schema_api_v1_platform_schema_center_label_schemas__key__patch",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelSchemaPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLabelSchemaRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Label Lab JSON schema",
        "tags": [
          "Schema Center"
        ]
      }
    },
    "/api/v1/platform/schema-center/label-schemas/{key}/archive": {
      "post": {
        "operationId": "api_schema_center_archive_label_schema_api_v1_platform_schema_center_label_schemas__key__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLabelSchemaRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Label Lab JSON schema",
        "tags": [
          "Schema Center"
        ]
      }
    },
    "/api/v1/platform/schema-center/label-schemas/{key}/purge": {
      "delete": {
        "operationId": "api_schema_center_purge_label_schema_api_v1_platform_schema_center_label_schemas__key__purge_delete",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purge archived Label Lab JSON schema",
        "tags": [
          "Schema Center"
        ]
      }
    },
    "/api/v1/platform/schema-center/label-schemas/{key}/restore": {
      "post": {
        "operationId": "api_schema_center_restore_label_schema_api_v1_platform_schema_center_label_schemas__key__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLabelSchemaRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore archived Label Lab JSON schema",
        "tags": [
          "Schema Center"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Agent Platform API local",
      "url": "http://127.0.0.1:8284"
    }
  ],
...
}